moved typedefs and resolvers into own file
